+++ lib/src/widgets/animated_svg_widget/animated_svg_widget_native.dart
@@ -42,7 +42,12 @@ class _AnimatedSvgWidgetState extends ConsumerState<AnimatedSvgWidget> {
       switch (htmlString) {
         case AsyncData(:final value):
           {
-            _webViewController.loadHtmlString(value);
+            // _webViewController.loadHtmlString(value);
+            if (widget.svgLink.startsWith("http")) {
+              _webViewController.loadRequest(Uri.parse(widget.svgLink));
+            } else {
+              _webViewController.loadFlutterAsset(widget.svgLink);
+            }
           }
       }
     });